chore: Move typing to folder types
