+++ @Reqeza/chunk-AI6V2MG2.mjs
@@ -1,0 +1,3 @@
+function f(e){return e==null||Object.keys(e).length===0}function d(e,r){let t=new Headers(e);if(r instanceof Headers)r.forEach((s,n)=>{t.append(n,s);});else if(typeof r=="object")for(let[s,n]of Object.entries(r))t.append(s,n);return t}var c=e=>/^[!#$%&'*+\-.^_`|~0-9a-zA-Z]+$/.test(e.trim()),a=e=>typeof e=="object"&&e!==null&&Object.keys(e).every(r=>typeof r=="string"&&c(r)&&typeof e[r]=="string"||Array.isArray(e[r])),p=e=>e.trim().toLowerCase().replace(/([a-z\d])(\w*)/g,(r,t,s)=>t.toUpperCase()+s),o=e=>e===!1||e==null?e:Array.isArray(e)?e.map(o):String(e),i=e=>Object.fromEntries(Object.entries(e).map(([r,t])=>[p(r),o(t)])),m=(e,r)=>{if(a(r)){let t=i(r);Object.assign(e,{...t});}else throw new TypeError("Invalid headers. Please provide valid headers.")};
+
+export { f as a, d as b, i as c, m as d };
