--- @Reqeza/chunk-D22QKJZO.js
@@ -1,3 +1,0 @@
-var d=Object.defineProperty;var e=(b,a,c)=>a in b?d(b,a,{enumerable:!0,configurable:!0,writable:!0,value:c}):b[a]=c;var f=(b,a,c)=>(e(b,typeof a!="symbol"?a+"":a,c),c);
-
-export { f as a };
