+++ @Reqeza/index.mjs
@@ -1,0 +1,7 @@
+import { b as b$1 } from './chunk-NARBQX72.mjs';
+import { d, a, b as b$3 } from './chunk-AI6V2MG2.mjs';
+import { b as b$2 } from './chunk-DCTSSTCD.mjs';
+
+var b=(s,e,n)=>{let r=new URL(e.path,s?.url??void 0);if(r.protocol!=="https:"&&r.protocol!=="http:")throw new TypeError(`Unsupported protocol, ${r.protocol}`);if(e.signal&&!(e.signal instanceof AbortSignal))throw new TypeError(typeof e.signal+" received for signal, but expected an AbortSignal");e?.qs&&(r.search=b$2.stringify(e.qs,s?.qs));let t=new Headers({...s?.headers});return e?.json&&(e.body=JSON.stringify(e.json),t.append("Content-Type","application/json; charset=UTF-8"),delete e.json),e?.body instanceof URLSearchParams&&t.append("Content-Type","application/x-www-form-urlencoded;charset=utf-8"),e.headers&&(t=b$3(t,e.headers)),new Request(r.toString(),{method:n.toLocaleUpperCase(),headers:t,body:e.body,signal:e.signal,cache:e.cache,credentials:e.credentials,integrity:e.integrity,keepalive:e.keepalive,mode:e.mode,redirect:e.redirect,referrer:e.referrer,referrerPolicy:e.referrerPolicy})},q=async(s,e,n)=>{let r=b(s,n,e);return s?.hooks&&await s.hooks.beforeRequest(r),fetch(r).then(t=>b$1(t,r,s)).then(async t=>{let a={headers:t.headers,status:t.status,statusText:t.statusText,config:r};if(n.responseType)try{a.data=await t[n.responseType]();}catch{throw new Error(`Unsupported response type "${n.responseType}" specified in the request. The Content-Type of the response is "${t.headers.get("Content-Type")}".`)}else a.data=await t.json();return a})},T=s=>{let e=["get","head","put","delete","post","patch","options"],n={json:"application/json",text:"text/*",formData:"multipart/form-data",arrayBuffer:"*/*",blob:"*/*"},r={};return e.forEach(t=>{r[t]=(a,o)=>{let d$1="json";if(typeof o?.auth=="object"){let p=`${o.auth.username}:${o.auth.password}`,u=Buffer.from(p).toString("base64");d(s?.headers,{Authorization:`Basic ${u}`});}let i={...Object.assign({},...Object.entries(n).map(([p,u])=>({[p]:()=>(d(s.headers=s?.headers||{},{accept:u}),d$1=p,i)}))),then(p){q(s,t,{path:a,responseType:d$1,...o}).then(p);}};return i};}),r},f=s=>{let e=Object.fromEntries(Object.entries(s||{default:{}}).map(([t,a])=>[t,{...T(a),beforeRequest:o=>{if(a?.hooks?.beforeRequest)throw new TypeError("beforeRequest has already been invoked within configuration.");a.hooks={},a.hooks.beforeRequest=o;},setHeaders:o=>d(a.headers=a.headers||{},o)}])),n=t=>{Object.keys(e).forEach(a=>{e[a].setHeaders(t);});};return a(s)?{...e.default}:{...e,...e[Object.keys(e)[0]],setHeaders:t=>n(t)}},O={create:f};
+
+export { O as default };
