--- @Reqeza/utils/errors.js
@@ -1,2 +1,0 @@
-export { a as HTTPError, b as ResponseError } from '../chunk-BKRRF6DW.js';
-import '../chunk-D22QKJZO.js';
