--- @Reqeza/utils/qs.js
@@ -1,2 +1,0 @@
-export { a as encode, b as qs } from '../chunk-45JH5YAD.js';
-import '../chunk-D22QKJZO.js';
