+++ package.json
@@ -2,8 +2,8 @@
   "name": "reqeza",
   "version": "2.2.7",
   "description": "Tiny JavaScript HTTP client based on browser Fetch API.",
-  "module": "@Reqeza/index.js",
-  "types": "@Reqeza/index.d.ts",
+  "module": "@Reqeza/index.mjs",
+  "types": "@Reqeza/index.d.mts",
   "scripts": {
     "build": "tsup",
     "test": "jest",
